updated distributed version to the latest version of twrp sequential
changes: Twrp: added the multimap support, updated the calculation of
hashmaps to stone only ids and cents instead of list of all cents,
updated the mergring of the hashmaps,  updated method to return the both
maps to return together.
Added the class ReturnMaps to enable returning 2 maps together.
Changed the Main spark class of RPhash to add this support for the new
twrp version changes.